refactor: ThemeManager 별도 파일로 분리, 테마 관련 파일 Theme 폴더로 정리, 주석 추가
